New JWT Authorization Grant Identity provider
Closes #43570
