+++ js/apps/admin-ui/maven-resources/theme/keycloak.v2/admin/messages/messages_en.properties
@@ -896,6 +896,7 @@ deleteError=Could not delete the provider {{error}}
 attributeDisplayName=Display name
 pkceEnabled=Use PKCE
 authorizationGrantSettings=Authorization Grant Settings
+authorizationGrantSettingsHelp= This section is used to configure specific settings related to JWT Authorization Grant support as defined in RFC 7523. Not all settings are configured here, some belong to the identity provider, such as JWKS URL, Issuer and Allowed Clock Skew.
 jwtAuthorizationGrantIdpEnabled=JWT Authorization Grant
 jwtAuthorizationGrantIdpEnabledHelp=Enable the identity provider to act as a trust provider to validate authorization grant JWT assertions according to RFC 7523.
 jwtAuthorizationGrantAssertionReuseAllowed=Allow assertion reuse
