+++ js/apps/admin-ui/src/identity-providers/add/DiscoverySettings.tsx
@@ -97,6 +97,7 @@ const Fields = ({ readOnly, isOIDC }: DiscoverySettingsProps) => {
               <DefaultSwitchControl
                 name="config.useJwksUrl"
                 label={t("useJwksUrl")}
+                labelIcon={t("useJwksUrlHelp")}
                 isDisabled={readOnly}
                 stringify
               />
